logistics-ai/backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.13.0
httpx>=0.27,<1.0
pymysql==1.1.0
sqlalchemy==2.0.23
bcrypt==4.1.1
pyjwt==2.12.1
cryptography==41.0.7
python-dotenv==1.0.0
email-validator==2.1.0
python-logging-loki
